Update README.md
Had to change reconfigure the new IP before reconnecting to webserver-2. Practiced with deleting .vsc-server using rm -rf to see if I could preserve the current working window. For the ansible host, I needed to open a new window as the current was missing the deleted file. At least I did not break a thing for now